+++ NEWS.md
@@ -5,7 +5,8 @@ ggptt 0.3.3.9003
 * `unset_ptt` to undo `set_ptt()`.
 * `set_gg()` and `unset_gg()` for general setting theme and scales
 * fixes to vingette.
-* small modifications to `theme_ptt()`
+* small modifications to `theme_ptt()`.
+* ... to `set_ptt()`.
 
 
 ggptt 0.3.3